+++ Tests~/Native/iOS/run-tests.sh
@@ -51,6 +51,7 @@ xcrun swiftc \
 
 EXPECTED_SYMBOLS=(
     AppstackUnityConfigure
+    AppstackUnitySetCustomerUserId
     AppstackUnitySendEvent
     AppstackUnityEnableAppleAdsAttribution
     AppstackUnityGetAppstackId